update to new breakout

+++ lib/Packet.cpp
@@ -9,7 +9,7 @@ Packet::Packet(uint32_t networkAddress, uint8_t hostAddressOctet, uint8_t client
     this->clientAddressOctet = clientAddressOctet;
     this->subDeviceID = subDeviceID;
     this->actionCode = actionCode;
-    for (int i = 0; i < ROIConstants::ROIMAXPACKETPAYLOAD && i < dataSize;
+    for (uint16_t i = 0; i < ROIConstants::ROIMAXPACKETPAYLOAD && i < dataSize;
          i++) {  // initialize data array
         this->data[i] = data[i];
     }
@@ -21,7 +21,7 @@ Packet::Packet(uint8_t hostAddressOctet, uint8_t clientAddressOctet) {
     this->clientAddressOctet = clientAddressOctet;
     this->subDeviceID = 0;
     this->actionCode = 0;
-    for (int i = 0; i < ROIConstants::ROIMAXPACKETPAYLOAD; i++) {  // initialize data array
+    for (uint16_t i = 0; i < ROIConstants::ROIMAXPACKETPAYLOAD; i++) {  // initialize data array
         this->data[i] = 0;
     }
 }
@@ -32,7 +32,7 @@ Packet::Packet() {
     this->clientAddressOctet = 0;
     this->subDeviceID = 0;
     this->actionCode = 0;
-    for (int i = 0; i < ROIConstants::ROIMAXPACKETPAYLOAD; i++) {  // initialize data array
+    for (uint16_t i = 0; i < ROIConstants::ROIMAXPACKETPAYLOAD; i++) {  // initialize data array
         this->data[i] = 0;
     }
 }
@@ -52,7 +52,7 @@ uint16_t Packet::getSubDeviceID() { return this->subDeviceID; }
 uint16_t Packet::getActionCode() { return this->actionCode; }
 
 void Packet::getData(uint8_t* dataBuffer, uint16_t dataBufferSize) {
-    for (int i = 0; i < ROIConstants::ROIMAXPACKETPAYLOAD && i < dataBufferSize; i++) {
+    for (uint16_t i = 0; i < ROIConstants::ROIMAXPACKETPAYLOAD && i < dataBufferSize; i++) {
         dataBuffer[i] = this->data[i];
     }
 }
@@ -72,7 +72,7 @@ void Packet::setSubDeviceID(uint16_t subDeviceID) { this->subDeviceID = subDevic
 void Packet::setActionCode(uint16_t actionCode) { this->actionCode = actionCode; }
 
 void Packet::setData(uint8_t* data, uint16_t dataSize) {
-    for (int i = 0; i < ROIConstants::ROIMAXPACKETPAYLOAD && i < dataSize;
+    for (uint16_t i = 0; i < ROIConstants::ROIMAXPACKETPAYLOAD && i < dataSize;
          i++) {  // initialize data array
         this->data[i] = data[i];
     }
@@ -84,7 +84,7 @@ bool Packet::importPacket(uint8_t* packet, uint16_t packetSize) {
     this->subDeviceID = (packet[0] << 8) | packet[1];
     this->actionCode = (packet[2] << 8) | packet[3];
     this->checksum = (packet[4] << 8) | packet[5];
-    for (int i = 6; i < ROIConstants::ROIMAXPACKETPAYLOAD + 6 && i < packetSize; i++) {
+    for (uint16_t i = 6; i < ROIConstants::ROIMAXPACKETPAYLOAD + 6 && i < packetSize; i++) {
         this->data[i - 6] = packet[i];  // copy data into data array
     }
     return validateChecksum();  // validate checksum
@@ -105,7 +105,7 @@ bool Packet::exportPacket(uint8_t* packetBuffer, uint16_t packetBufferSize) {
     packetBuffer[4] = (this->checksum >> 8) & 0xff;
     packetBuffer[5] = this->checksum & 0xff;
 
-    for (int i = 0; i < ROIConstants::ROIMAXPACKETPAYLOAD && i < packetBufferSize - 6; i++) {
+    for (uint16_t i = 0; i < ROIConstants::ROIMAXPACKETPAYLOAD && i < packetBufferSize - 6; i++) {
         packetBuffer[i + 6] = this->data[i];
     }
     return true;
@@ -121,7 +121,7 @@ uint16_t Packet::calculateChecksum() {
     // retried.
     checksum += this->subDeviceID;
     checksum += this->actionCode;
-    for (int i = 0; i < ROIConstants::ROIMAXPACKETPAYLOAD; i++) {
+    for (uint16_t i = 0; i < ROIConstants::ROIMAXPACKETPAYLOAD; i++) {
         checksum += this->data[i];
     }
 
@@ -143,7 +143,7 @@ sysAdminPacket::sysAdminPacket(uint32_t networkAddress, uint8_t hostAddressOctet
     this->clientAddressOctet = clientAddressOctet;
     this->subDeviceID = subDeviceID;
     this->actionCode = actionCode;
-    for (int i = 0; i < ROIConstants::ROIMAXPACKETPAYLOAD && i < dataBufferSize;
+    for (uint16_t i = 0; i < ROIConstants::ROIMAXPACKETPAYLOAD && i < dataBufferSize;
          i++) {  // initialize data array
         this->data[i] = data[i];
     }
@@ -157,7 +157,7 @@ sysAdminPacket::sysAdminPacket(uint8_t hostAddressOctet, uint8_t clientAddressOc
     this->clientAddressOctet = clientAddressOctet;
     this->subDeviceID = 0;
     this->actionCode = 0;
-    for (int i = 0; i < ROIConstants::ROIMAXPACKETPAYLOAD; i++) {  // initialize data array
+    for (uint16_t i = 0; i < ROIConstants::ROIMAXPACKETPAYLOAD; i++) {  // initialize data array
         this->data[i] = 0;
     }
     this->adminMetaData = 0;
@@ -170,7 +170,7 @@ sysAdminPacket::sysAdminPacket() {
     this->clientAddressOctet = 0;
     this->subDeviceID = 0;
     this->actionCode = 0;
-    for (int i = 0; i < ROIConstants::ROIMAXPACKETPAYLOAD; i++) {  // initialize data array
+    for (uint16_t i = 0; i < ROIConstants::ROIMAXPACKETPAYLOAD; i++) {  // initialize data array
         this->data[i] = 0;
     }
     this->adminMetaData = 0;
@@ -203,7 +203,7 @@ bool sysAdminPacket::importPacket(uint8_t* packet, uint16_t packetSize) {
     this->originHostOctet = packet[2];
     this->actionCode = (packet[3] << 8) | packet[4];
     this->checksum = (packet[5] << 8) | packet[6];
-    for (int i = 7; i < ROIConstants::ROIMAXPACKETPAYLOAD + 7 && i < packetSize;
+    for (uint16_t i = 7; i < ROIConstants::ROIMAXPACKETPAYLOAD + 7 && i < packetSize;
          i++) {  // initialize data array
         this->data[i - 7] = packet[i];
     }
@@ -231,7 +231,7 @@ bool sysAdminPacket::exportPacket(uint8_t* packetBuffer, uint16_t packetBufferSi
     packetBuffer[4] = this->actionCode & 0xff;
     packetBuffer[5] = (this->checksum >> 8) & 0xff;
     packetBuffer[6] = this->checksum & 0xff;
-    for (int i = 0; i < ROIConstants::ROIMAXPACKETPAYLOAD && i < packetBufferSize - 7; i++) {
+    for (uint16_t i = 0; i < ROIConstants::ROIMAXPACKETPAYLOAD && i < packetBufferSize - 7; i++) {
         packetBuffer[i + 7] = this->data[i];
     }
     return true;
@@ -248,7 +248,7 @@ uint16_t sysAdminPacket::calculateChecksum() {
     checksum += this->adminMetaData;
     checksum += this->originHostOctet;
     checksum += this->actionCode;
-    for (int i = 0; i < ROIConstants::ROIMAXPACKETPAYLOAD; i++) {
+    for (uint16_t i = 0; i < ROIConstants::ROIMAXPACKETPAYLOAD; i++) {
         checksum += this->data[i];
     }
